Fixes build issue in phipshape image.
In the previous campfire build files the paths to files
were from the location of the build file, while with bazel
the root is from the root of the repo.

This build target must be manually triggered and hasn't caused
trouble until now, when a release (building this image and others)
was finally done again.